Core update - Updated Json library to latest version and .NET 4.5 - Added Localization support - Fixed a bug in the CommandLine parser causing short arguments to fail - added helper to reverse a string - added AddRange for dictionaries

+++ UtilsCommandLine/Logic/CommandLineArguments.cs
@@ -263,8 +263,8 @@
 
             foreach (ICommandLineSwitchDefinition definition in this.definitions)
             {
-                ICommandLineSwitch activeLong = this.GetSwitch(definition.Short);
-                ICommandLineSwitch activeShort = this.GetSwitch(definition.Long);
+                ICommandLineSwitch activeLong = this.GetSwitch(definition.Long);
+                ICommandLineSwitch activeShort = this.GetSwitch(definition.Short);
                 if (activeLong == null && activeShort == null)
                 {
                     if (definition.Required)
